Revert "ENH: stats.ks_2samp: Pythranize remaining exact p-value calculations (#13957)"
This reverts commit aac2870bfbfa7047700fcc89244e89e409b2461d.
