+++ scipy/stats/setup.py
@@ -47,6 +47,14 @@ def configuration(parent_package='', top_path=None):
                                sources=['_qmc_cy.cxx'])
     ext._pre_build_hook = set_cxx_flags_hook
 
+    if int(os.environ.get('SCIPY_USE_PYTHRAN', 1)):
+        import pythran
+        ext = pythran.dist.PythranExtension(
+            'scipy.stats._stats_pythran',
+            sources=["scipy/stats/_stats_pythran.py"],
+            config=['compiler.blas=none'])
+        config.ext_modules.append(ext)
+
     # add BiasedUrn module
     config.add_data_files('_biasedurn.pxd')
     from _generate_pyx import isNPY_OLD  # type: ignore[import]
